requirements.txt
torchvision>=0.14.0
timm~=0.4.9
numpy>=1.23.4
scipy>=1.9.3
imgaug>=0.4.0
pillow>=9.2.0
matplotlib>=3.5.3
hub>=3.0.1
pandas>=1.5.1
